+++ packages/app/src/cli/services/dev/theme-extension-args.test.ts
@@ -1,7 +1,6 @@
 import {themeExtensionArgs} from './theme-extension-args'
-import {AppInterface} from '../../models/app/app'
-import {ThemeExtension} from '../../models/app/extensions'
 import {ensureDeployEnvironment} from '../environment'
+import {testApp, testThemeExtensions} from '../../models/app/app.test-data'
 import {beforeAll, describe, expect, it, vi} from 'vitest'
 
 beforeAll(() => {
@@ -12,9 +11,9 @@ beforeAll(() => {
 describe('themeExtensionArgs', async () => {
   it('returns valid theme extension arguments', async () => {
     const apiKey = 'api_key_0000_1111_2222'
-    const app = testApp()
     const reset = false
-    const options = {app, reset, port: 8282, theme: 'theme ID'}
+    const options = {app: testApp(), reset, themeExtensionPort: 8282, theme: 'theme ID'}
+    const extension = testThemeExtensions()
 
     const deployEnvironmentOutput = {
       app: options.app,
@@ -36,7 +35,7 @@ describe('themeExtensionArgs', async () => {
 
     vi.mocked(ensureDeployEnvironment).mockReturnValue(Promise.resolve(deployEnvironmentOutput))
 
-    const args = await themeExtensionArgs(apiKey, options)
+    const args = await themeExtensionArgs(extension, apiKey, options)
 
     expect(args).toEqual([
       './my-extension',
@@ -59,38 +58,3 @@ describe('themeExtensionArgs', async () => {
     ])
   })
 })
-
-export function testApp(): AppInterface {
-  return {
-    name: 'App',
-    idEnvironmentVariableName: 'SHOPIFY_API_KEY',
-    directory: '/tmp/project',
-    packageManager: 'yarn',
-    configuration: {scopes: ''},
-    configurationPath: '/tmp/project/shopify.app.toml',
-    nodeDependencies: {},
-    webs: [],
-    extensions: {
-      ui: [],
-      theme: [testThemeExtensions()],
-      function: [],
-    },
-    hasExtensions: () => true,
-    updateDependencies: async () => {},
-  }
-}
-
-export function testThemeExtensions(): ThemeExtension {
-  return {
-    configuration: {
-      name: 'theme extension name',
-      type: 'theme',
-    },
-    idEnvironmentVariableName: '',
-    localIdentifier: 'extension title',
-    configurationPath: '',
-    directory: './my-extension',
-    type: 'theme',
-    graphQLType: 'THEME_APP_EXTENSION',
-  }
-}
